[Data rearchitecture] Add `CustomRevisionFilter` module to shared logic for filtering revisions. (#6219)
* Add CustomRevisionFilter module to handle shared logic between ArticleScopedProgram and VisitingScholarship.
