+++ app/models/course_types/visiting_scholarship.rb
@@ -50,10 +50,12 @@
 #
 
 class VisitingScholarship < Course
+  include CustomRevisionFilter
+
   has_many(:revisions, lambda do |course|
     where('date >= ?', course.start)
     .where('date <= ?', course.end)
-    .where(article_id: course.assignments.pluck(:article_id))
+    .where(article_id: course.assigned_article_ids)
   end, through: :students)
 
   def wiki_edits_enabled?
@@ -76,20 +78,7 @@ class VisitingScholarship < Course
     true
   end
 
-  def filter_revisions(wiki, revisions)
-    filtered_data = revisions.select do |_, details|
-      article_title = details['article']['title']
-      formatted_article_title = ArticleUtils.format_article_title(article_title, wiki)
-      scoped_article_titles.include?(formatted_article_title)
-    end
-    filtered_data
-  end
-
   def scoped_article_titles
-    assignments.pluck(:article_title)
-  end
-
-  def scoped_article_ids
-    assignments.pluck(:article_id)
+    assigned_article_titles
   end
 end
